requirements.txt
langchain>=0.1.0
langchain-ollama>=0.0.1
langchain-core>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.0.1
chromadb>=0.4.0
mysql-connector-python>=8.0.0
toolbox-langchain>=0.1.0
gradio>=4.0.0
pyyaml>=6.0
python-dateutil>=2.8.2
requests>=2.28.0